Works on CharacterCreator
Added a new int value to Ability class named acquireCost
This value specifies the number of ability points needed to acquire the
ability
Updated the Ability constructor to initialize new value
Changed the skillPoints and abilityPoints variables inside Character to
private
Added the following functions to view and alter skillPoints and
abilityPoints:
hasPoints
getPoints
addPoints
subPoints
Added new function to CharacterCreator to allow adding new abilities
Added tryParse method to Helper class. May be further implemented later,
but will probably end up being reverted due to redundancy
Corrected the Log function in DebugLog so that it correctly takes in an
action and a message
The function also only redirects if debugging is enabled.
Added function invalidInputError that is used to specify invalid inputs
Added Error function to output errors to the console