Effect Class
Created class for effects.
Effects are applied to character to effect stats
Created class for listing and initializing effects
Added and implemented the ChangeSkill effect.
This effect allows the draining/fortifying of any skill/stat the player
has
Made the StatViewer a static class
Removed the skillEnhance variable from Ability and replaced it with an
array for effects.
This array tracks what effects an ability has and applies those effects
to the Ability's target
Added a target reference to Ability
Added a Cast function to ActionHandler.
This allows the user to cast abilities
Added a list of effects to the character as well as a list of effects to
remove.
After every action, effects take place until their duration runs out, at
which point they are added to removeEffects and removed from Effects
Added functions to Character for updating and removing effects
Removed the assigning of Fireball in AbilityList from its declaration to
the InitializeAbilityList method.