Implementing CharacterCreator
Implemented the distributePoints methods in Character Creator
These allow the user to distribute their skill and ability points to
upgrade skills and abilities
AddAbility method allows users to acquire new abilities
Added a boolean parameter for processChoice in Helper class
This allows for easier iteration through both lists whose inputs start
at 0 or 1 with a simple change of a value
Added viewAbilityArray() to StatViewer
This allows the passing of a custom ability array to view in the same
format as ViewAbilityList
Character is now initialized in MainMenu
Can view character stats and abilities from main menu
Can spend ability and skill points from their respective view stat
screens
Added boolean functions for checking manaCost, healthCost and
buildUpCost
These fuctions will prevent these values from falling below 0
The field strength in Spells is now a float. Functions that make use of
it will now use Math.Floor casted to an int for the correct value
This change is so that xp can modify the value to a more precise level
Cast method in ActionHandler now takes in a second Character parameter
for target. This allows easier access to the Target field in the Ability
class
Character class now initializes the character