Continuing the work on battle
Added new function calcReaction to ActionHandler
This method returns if an entity is given a reaction
Added function to Battle getPositionToPlayer
This returns the position of enemies in relation to the player for
informative output
Added combatCycle function to Battle
Controls the overall cycle of the battle. Runs so long as the battle is
active
Added combatPhases method to battle
Controls the Action cycle specifically. Runs until the next character's
turn
Added determineOrder method to Battle
Determines order of the characters based on highest initiative
Added calcReactions to Battle to ensure that the method in ActionHandler
gets called for every entity that needs to be checked.
All other methods added to Battle will likely be removed
Changed xPosition and yPosition to ints in Character
Added a new int field: actions.
This value shows how many actions a character has remaining on their
turn
Added new boolean field canReact
Holds a value for if a character is granted a reaction
Added calculateActions method to Helper class
This method determines how many actions a character gets on a given
turn.
Additionally added a getChoice method. This method takes in a message
and options and then outputs them in the console so that the user can
make choices.
This should cut down on the need for a "displayThis" method every time
the game requires input.