Probability Helper
Added new class: ProbabilityHelper
Helps determine various AI actions
Added new values to Ability:
range - the distance an ability can travel
probability - the chance this ability is given to an AI
Created 3 enums in AI class:
Title - Various titles that can be given to classify AI and increase
probabilities in obtaining certain abilities and skills and performing
certain actions
Action - The Actions an AI can perform
Goals - The goal of the AI. This influences the actions that the AI
takes
Added healthStatus enum to Character. This will be used to inform users
on an enemy's health and will also be used to increase the probability
of an AI attempting a heal provided that they have one.
Added new method calcStatus to Character. Function assigns a health
status based on the entity's current health