Items, Calculations, and Furthering Battle
Created a calculations class where all calculations are done
Removed all calculations from ActionHandler to Calculations
Added functions to allow characters to move diagonally in ActionHandler
Added function to arena  to allow placing of characters
Made the initiating character a global variable that changes every turn
Added functions to allow moving should that be the chosen action to
Battle
Added function to select a target if attacking or casting
Added a function calcMeleeDamage to calculate melee damage
Reduced the amount of Goals and Actions in AI so that the process can be
better streamlined
Added a new class Item and ItemList
Items are used by characters to increase their stats.
ItemList is where all items that will reside in the game are created
Added List of Items variable to Character to store the items the
character has
Added an addItem function to add items and apply their stats to
character
Added a removeItem function to add items and remove their stats from
character
getCharacterSkillList in Helper now takes in an Object instead of a
Character. This allows items to use this function as well
Added a shuffle method to character class which shuffles any list of
values
Added new int value "range" to Object. This value determines how far an
object can reach.
Added new Skill "defense". This is used to deflect incoming damage